requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.21.0
asyncpgsa==0.26.3
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
emoji==1.2.0
environs==8.0.0
greenlet==1.0.0
hiredis==1.1.0
idna==3.1
marshmallow==3.10.0
multidict==5.1.0
python-dotenv==0.15.0
pytz==2021.1
SQLAlchemy==1.4.0b2
typing-extensions==3.7.4.3
yarl==1.6.3
tortoise-orm==0.16.21
anyio==2.1.0
APScheduler==3.7.0